scripts/requirements.txt
transformers>=4.40.0
torch>=2.3.0
peft>=0.11.0
datasets>=2.20.0
accelerate>=0.30.0
scipy>=1.13.0
scikit-learn>=1.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
